backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-cloud-discoveryengine==0.13.4
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.2